Updating All code to update dynamically.
Fixed a lot of bugs, the that was occurred when copied code in (12. Updating All code to update dynamically). did not work. The array [] is not being able to add in inline styles, adding ...styles is giving error. had to add inline styles as an object {{  }}, in gameOverScreen I used Object.assign({}, style1, style2) to add an inline style with imageContainer Style to make it dynamic.